requirements.txt
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
click==6.7
darkskylib==0.3.91
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
future==0.18.2
idna==2.8
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
psycopg2-binary==2.7.5
requests==2.22.0
SQLAlchemy==1.2.8
urllib3==1.25.7
Werkzeug==0.14.1
twilio~=6.0.0
